requirements.txt
python-telegram-bot==21.10
SQLAlchemy==2.0.20
psycopg2-binary==2.9.10
python-dotenv==1.0.0
pydantic_settings==2.7.1
loguru
asyncpg==0.30.0
greenlet==3.1.1
opencv-python==4.10.0.84



